Refactor request logging
